ml-risk-service/requirements.txt
# Core ML dependencies
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Ranking algorithms
scikit-multilearn==0.2.0
lightgbm==4.0.0

# Explainability
shap==0.42.1

# gRPC and API
grpcio==1.57.0
grpcio-tools==1.57.0
protobuf==4.24.0

# Serialization and data handling
joblib==1.3.1
pyyaml==6.0.1
jsonschema==4.19.0

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1